chore(upgrade): move everlight upgrade target from v1.15.0 to v1.12.0
