+++ app/upgrades/upgrades.go
@@ -15,7 +15,7 @@ import (
 	upgrade_v1_10_1 "github.com/LumeraProtocol/lumera/app/upgrades/v1_10_1"
 	upgrade_v1_11_0 "github.com/LumeraProtocol/lumera/app/upgrades/v1_11_0"
 	upgrade_v1_11_1 "github.com/LumeraProtocol/lumera/app/upgrades/v1_11_1"
-	upgrade_v1_15_0 "github.com/LumeraProtocol/lumera/app/upgrades/v1_15_0"
+	upgrade_v1_12_0 "github.com/LumeraProtocol/lumera/app/upgrades/v1_12_0"
 	upgrade_v1_6_1 "github.com/LumeraProtocol/lumera/app/upgrades/v1_6_1"
 	upgrade_v1_8_0 "github.com/LumeraProtocol/lumera/app/upgrades/v1_8_0"
 	upgrade_v1_8_4 "github.com/LumeraProtocol/lumera/app/upgrades/v1_8_4"
@@ -38,7 +38,7 @@ import (
 // | v1.10.1 | custom   | drop crisis (if not already)      | Ensure consensus params are present in x/consensus
 // | v1.11.0 | custom   | add audit store                   | Initializes audit params with dynamic epoch_zero_height
 // | v1.11.1 | custom   | conditional add audit store       | Supports direct v1.10.1->v1.11.1 and enforces audit min_disk_free_percent floor (>=15)
-// | v1.15.0 | custom   | none (Everlight in supernode)     | Runs migrations; Everlight logic embedded in x/supernode
+// | v1.12.0 | custom   | none (Everlight in supernode)     | Runs migrations; Everlight logic embedded in x/supernode
 // =================================================================================================================================
 
 type UpgradeConfig struct {
@@ -68,7 +68,7 @@ var upgradeNames = []string{
 	upgrade_v1_10_1.UpgradeName,
 	upgrade_v1_11_0.UpgradeName,
 	upgrade_v1_11_1.UpgradeName,
-	upgrade_v1_15_0.UpgradeName,
+	upgrade_v1_12_0.UpgradeName,
 }
 
 var NoUpgradeConfig = UpgradeConfig{
@@ -143,10 +143,10 @@ func SetupUpgrades(upgradeName string, params appParams.AppUpgradeParams) (Upgra
 			StoreUpgrade: &upgrade_v1_11_1.StoreUpgrades,
 			Handler:      upgrade_v1_11_1.CreateUpgradeHandler(params),
 		}, true
-	case upgrade_v1_15_0.UpgradeName:
+	case upgrade_v1_12_0.UpgradeName:
 		return UpgradeConfig{
-			StoreUpgrade: &upgrade_v1_15_0.StoreUpgrades,
-			Handler:      upgrade_v1_15_0.CreateUpgradeHandler(params),
+			StoreUpgrade: &upgrade_v1_12_0.StoreUpgrades,
+			Handler:      upgrade_v1_12_0.CreateUpgradeHandler(params),
 		}, true
 
 	// add future upgrades here
