+++ app/upgrades/v1_12_0/store.go
@@ -1,9 +1,9 @@
-package v1_15_0
+package v1_12_0
 
 import (
 	storetypes "cosmossdk.io/store/types"
 )
 
-// StoreUpgrades defines store changes for v1.15.0.
+// StoreUpgrades defines store changes for v1.12.0.
 // Everlight now lives inside x/supernode, so no dedicated store is added.
 var StoreUpgrades = storetypes.StoreUpgrades{}
