+++ app/upgrades/v1_12_0/upgrade.go
@@ -1,4 +1,4 @@
-package v1_15_0
+package v1_12_0
 
 import (
 	"context"
@@ -14,7 +14,7 @@ import (
 )
 
 // UpgradeName is the on-chain name used for this upgrade.
-const UpgradeName = "v1.15.0"
+const UpgradeName = "v1.12.0"
 
 // CreateUpgradeHandler runs the migration that initializes the supernode
 // module's embedded Everlight configuration.
