+++ app/upgrades/v1_12_0/upgrade_test.go
@@ -1,4 +1,4 @@
-package v1_15_0
+package v1_12_0
 
 import (
 	"regexp"
